backend/requirements.txt
fastapi 
uvicorn
pydantic
sqlalchemy
pymongo
motor
mysql-connector-python
dotenv
